ppc_dashboard/backend/requirements.txt
Flask==3.1.3
flask_cors==6.0.2
nncf==3.0.0
numpy==2.4.2
onnxruntime_gpu==1.23.2
opencv_python==4.12.0.88
psutil==7.0.0
PyYAML==6.0.2
PyYAML==6.0.3
timm==1.0.25
torch==2.10.0+cu128
transformers==4.57.1
ultralytics==8.3.226
waitress==3.0.2
